Move drive with align out of basic commands
