+++ src/main/java/frc/Board/ErrorTab.java
@@ -1,7 +1,6 @@
 // Copyright (c) FIRST and other WPILib contributors.
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
-
 package frc.Board;
 
 import edu.wpi.first.networktables.GenericEntry;
@@ -9,11 +8,7 @@ import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 
 public class ErrorTab {
-
   Shuffleboard m_sb;
-
-  // Shuffleboard DriveTrain entries
-  // Create and get reference to SB tab
 
   ShuffleboardTab m_sbt_Errors;
 
@@ -37,11 +32,10 @@ public class ErrorTab {
   }
 
   private void initShuffleboardTab() {
-    // Create and get reference to SB tab
+    // Create and get reference to the tab
     m_sbt_Errors = Shuffleboard.getTab("Errors");
 
     m_nte_Errors = m_sbt_Errors.add("Errors", CurrentErrors).getEntry();
-    // Create widgets for PID Controllers
   }
 
 
@@ -51,7 +45,7 @@ public class ErrorTab {
       m_nte_Errors = m_sbt_Errors.add(nameOfCaller, newError).getEntry();
     }
   }
-  public boolean errorExists(String nameOfCaller, String newError){
+  public boolean errorExists(String nameOfCaller, String newError) {
     if(CurrentErrors.contains(newError)){
       return true;
     }else{
