+++ src/main/java/frc/robot/RobotContainer.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 
 import frc.robot.constants.OIConstants;
 import frc.robot.commands.AlignToTagPhotonVision;
-import frc.robot.commands.basic.DriveWhileAligning;
+import frc.robot.commands.DriveWhileAligning;
 import frc.robot.subsystems.DriveSubsystem;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
